Reorganize Go tests and add CI
