+++ unit_test.go
@@ -2,17 +2,26 @@ package main
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
+	"fmt"
 	"html/template"
+	"log"
+	"net/http"
+	"net/http/httptest"
+	"os"
 	"strings"
+	"sync"
+	"sync/atomic"
 	"testing"
+	"time"
 
 	"github.com/PuerkitoBio/goquery"
 )
 
 // ---------------- Rendering: formatting and message rendering ----------------
 
-func TestMessageFormatting(t *testing.T) {
+func TestUnitMessageFormatting(t *testing.T) {
 	templates, err := loadTemplates()
 	if err != nil {
 		t.Fatalf("Failed to load templates: %v", err)
@@ -76,7 +85,7 @@ func TestMessageFormatting(t *testing.T) {
 	}
 }
 
-func TestUserAndLLMRenderingSame(t *testing.T) {
+func TestUnitUserAndLLMRenderingSame(t *testing.T) {
 	templates, err := loadTemplates()
 	if err != nil {
 		t.Fatalf("Failed to load templates: %v", err)
@@ -101,7 +110,7 @@ func TestUserAndLLMRenderingSame(t *testing.T) {
 	}
 }
 
-func TestMultilineInputSupport(t *testing.T) {
+func TestUnitMultilineInputSupport(t *testing.T) {
 	templates, err := loadTemplates()
 	if err != nil {
 		t.Fatalf("Failed to load templates: %v", err)
@@ -138,7 +147,7 @@ func TestMultilineInputSupport(t *testing.T) {
 
 // ---------------- Rendering: template auto-escaping and newline preservation ----------------
 
-func TestHTMLEscapingInTemplates(t *testing.T) {
+func TestUnitHTMLEscapingInTemplates(t *testing.T) {
 	tmpl := template.Must(template.New("test").Parse(`<div>{{.Content}}</div>`))
 	tests := []struct{ name, input, expected string }{
 		{"Script tag", "<script>alert('XSS')</script>", "<div>&lt;script&gt;alert(&#39;XSS&#39;)&lt;/script&gt;</div>"},
@@ -166,7 +175,7 @@ func TestHTMLEscapingInTemplates(t *testing.T) {
 	}
 }
 
-func TestNewlinePreservation(t *testing.T) {
+func TestUnitNewlinePreservation(t *testing.T) {
 	tmpl := template.Must(template.New("test").Parse(`<div class="preserve-breaks">{{.}}</div>`))
 	tests := []struct {
 		name, input string
@@ -196,7 +205,7 @@ func TestNewlinePreservation(t *testing.T) {
 	}
 }
 
-func TestMessagePartDataSecurity(t *testing.T) {
+func TestUnitMessagePartDataSecurity(t *testing.T) {
 	tests := []struct {
 		name, partType, content string
 		isMarkdown              bool
@@ -224,7 +233,7 @@ func TestMessagePartDataSecurity(t *testing.T) {
 
 // ---------------- Rendering: HTTP part transformation and TODO renderer ----------------
 
-// Helpers from unit_http_parts_test.go
+// Helpers consolidated from the previous unit_http_parts_test.go
 func convertToMessagePart(enhanced EnhancedMessagePart) MessagePart {
 	return MessagePart{
 		ID:        enhanced.ID,
@@ -303,7 +312,7 @@ func padLeft(s string, length int, pad rune) string {
 	return s
 }
 
-func TestTransformTextPart(t *testing.T) {
+func TestUnitTransformTextPart(t *testing.T) {
 	templates, err := loadTemplates()
 	if err != nil {
 		t.Fatal(err)
@@ -328,7 +337,7 @@ func TestTransformTextPart(t *testing.T) {
 	}
 }
 
-func TestTransformToolPart(t *testing.T) {
+func TestUnitTransformToolPart(t *testing.T) {
 	templates, err := loadTemplates()
 	if err != nil {
 		t.Fatal(err)
@@ -350,7 +359,7 @@ func TestTransformToolPart(t *testing.T) {
 	}
 }
 
-func TestTransformTodoWritePart(t *testing.T) {
+func TestUnitTransformTodoWritePart(t *testing.T) {
 	templates, err := loadTemplates()
 	if err != nil {
 		t.Fatal(err)
@@ -370,7 +379,7 @@ func TestTransformTodoWritePart(t *testing.T) {
 	}
 }
 
-func TestTransformReasoningPart(t *testing.T) {
+func TestUnitTransformReasoningPart(t *testing.T) {
 	templates, err := loadTemplates()
 	if err != nil {
 		t.Fatal(err)
@@ -388,7 +397,7 @@ func TestTransformReasoningPart(t *testing.T) {
 	}
 }
 
-func TestTransformStepParts(t *testing.T) {
+func TestUnitTransformStepParts(t *testing.T) {
 	templates, err := loadTemplates()
 	if err != nil {
 		t.Fatal(err)
@@ -417,7 +426,7 @@ func TestTransformStepParts(t *testing.T) {
 	}
 }
 
-func TestTransformMessageWithAllParts(t *testing.T) {
+func TestUnitTransformMessageWithAllParts(t *testing.T) {
 	templates, err := loadTemplates()
 	if err != nil {
 		t.Fatal(err)
@@ -466,7 +475,7 @@ func TestTransformMessageWithAllParts(t *testing.T) {
 
 // ---------------- Rendering: templates and UI ----------------
 
-func TestTemplateConsistency(t *testing.T) {
+func TestUnitTemplateConsistency(t *testing.T) {
 	templates, err := loadTemplates()
 	if err != nil {
 		t.Fatal(err)
@@ -587,7 +596,7 @@ func TestTemplateConsistency(t *testing.T) {
 	})
 }
 
-func TestTemplateHTMLStructure(t *testing.T) {
+func TestUnitTemplateHTMLStructure(t *testing.T) {
 	templates, err := loadTemplates()
 	if err != nil {
 		t.Fatal(err)
@@ -635,7 +644,7 @@ func TestTemplateHTMLStructure(t *testing.T) {
 	})
 }
 
-func TestMessageTemplateIDs(t *testing.T) {
+func TestUnitMessageTemplateIDs(t *testing.T) {
 	templates, err := loadTemplates()
 	if err != nil {
 		t.Fatal(err)
@@ -698,7 +707,7 @@ func TestMessageTemplateIDs(t *testing.T) {
 	}
 }
 
-func TestMessageMetadata(t *testing.T) {
+func TestUnitMessageMetadata(t *testing.T) {
 	templates, err := loadTemplates()
 	if err != nil {
 		t.Fatal(err)
@@ -741,7 +750,7 @@ func TestMessageMetadata(t *testing.T) {
 	}
 }
 
-func TestStreamingCSS(t *testing.T) {
+func TestUnitStreamingCSS(t *testing.T) {
 	cssBytes, err := staticFS.ReadFile("static/styles.css")
 	if err != nil {
 		t.Fatal(err)
@@ -758,7 +767,7 @@ func TestStreamingCSS(t *testing.T) {
 	}
 }
 
-func TestFileDropdownSelectionPreservation(t *testing.T) {
+func TestUnitFileDropdownSelectionPreservation(t *testing.T) {
 	templates, err := loadTemplates()
 	if err != nil {
 		t.Fatal(err)
@@ -827,7 +836,7 @@ func TestFileDropdownSelectionPreservation(t *testing.T) {
 
 // ---------------- Rendering: Todo list renderer ----------------
 
-func TestRenderTodoList(t *testing.T) {
+func TestUnitRenderTodoList(t *testing.T) {
 	templates, err := loadTemplates()
 	if err != nil {
 		t.Fatalf("Failed to load templates: %v", err)
@@ -857,5 +866,834 @@ func TestRenderTodoList(t *testing.T) {
 				}
 			}
 		})
+	}
+}
+
+// ---------------- Server: logging middleware/response writer ----------------
+
+func TestUnitLoggingResponseWriter_WriteHeader(t *testing.T) {
+	recorder := httptest.NewRecorder()
+	lw := NewLoggingResponseWriter(recorder)
+	lw.WriteHeader(404)
+	if lw.statusCode != 404 {
+		t.Errorf("Expected status code 404, got %d", lw.statusCode)
+	}
+	if recorder.Code != 404 {
+		t.Errorf("Expected recorder status code 404, got %d", recorder.Code)
+	}
+}
+
+func TestUnitLoggingResponseWriter_Write(t *testing.T) {
+	recorder := httptest.NewRecorder()
+	lw := NewLoggingResponseWriter(recorder)
+	testData := []byte("Hello, World!")
+	n, err := lw.Write(testData)
+	if err != nil {
+		t.Errorf("Write returned error: %v", err)
+	}
+	if n != len(testData) {
+		t.Errorf("Expected to write %d bytes, wrote %d", len(testData), n)
+	}
+	if lw.body.String() != "Hello, World!" {
+		t.Errorf("Expected body to contain 'Hello, World!', got '%s'", lw.body.String())
+	}
+	if recorder.Body.String() != "Hello, World!" {
+		t.Errorf("Expected recorder body to contain 'Hello, World!', got '%s'", recorder.Body.String())
+	}
+}
+
+func TestUnitLoggingResponseWriter_LogResponse(t *testing.T) {
+	var logOutput bytes.Buffer
+	log.SetOutput(&logOutput)
+	defer log.SetOutput(os.Stderr)
+	recorder := httptest.NewRecorder()
+	lw := NewLoggingResponseWriter(recorder)
+	lw.WriteHeader(200)
+	if _, err := lw.Write([]byte("<html><body>Test Response</body></html>")); err != nil {
+		t.Fatalf("Failed to write: %v", err)
+	}
+	lw.LogResponse("GET", "/test")
+	logStr := logOutput.String()
+	if !strings.Contains(logStr, "WIRE_OUT GET /test [200]") {
+		t.Errorf("Expected log to contain 'WIRE_OUT GET /test [200]', got: %s", logStr)
+	}
+	if !strings.Contains(logStr, "<html><body>Test Response</body></html>") {
+		t.Errorf("Expected log to contain full response body, got: %s", logStr)
+	}
+}
+
+func TestUnitLoggingResponseWriter_LogResponseNoTruncation(t *testing.T) {
+	var logOutput bytes.Buffer
+	log.SetOutput(&logOutput)
+	defer log.SetOutput(os.Stderr)
+	recorder := httptest.NewRecorder()
+	lw := NewLoggingResponseWriter(recorder)
+	largeBody := strings.Repeat("A", 1000)
+	if _, err := lw.Write([]byte(largeBody)); err != nil {
+		t.Fatalf("Failed to write large body: %v", err)
+	}
+	lw.LogResponse("POST", "/large")
+	logStr := logOutput.String()
+	if !strings.Contains(logStr, largeBody) {
+		t.Errorf("Expected log to contain full large body without truncation")
+	}
+	if strings.Contains(logStr, "truncated") {
+		t.Errorf("Log should not contain truncation message, but does: %s", logStr)
+	}
+}
+
+func TestUnitLoggingMiddleware_NormalEndpoint(t *testing.T) {
+	var logOutput bytes.Buffer
+	log.SetOutput(&logOutput)
+	defer log.SetOutput(os.Stderr)
+	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.WriteHeader(200)
+		w.Write([]byte("Normal response"))
+	})
+	req := httptest.NewRequest("GET", "/normal", nil)
+	recorder := httptest.NewRecorder()
+	wrappedHandler := loggingMiddleware(handler)
+	wrappedHandler.ServeHTTP(recorder, req)
+	if recorder.Code != 200 {
+		t.Errorf("Expected status code 200, got %d", recorder.Code)
+	}
+	if recorder.Body.String() != "Normal response" {
+		t.Errorf("Expected 'Normal response', got '%s'", recorder.Body.String())
+	}
+	if !strings.Contains(logOutput.String(), "WIRE_OUT GET /normal [200]: Normal response") {
+		t.Errorf("Expected log output for normal endpoint, got: %s", logOutput.String())
+	}
+}
+
+func TestUnitLoggingMiddleware_SSEEndpoint(t *testing.T) {
+	var logOutput bytes.Buffer
+	log.SetOutput(&logOutput)
+	defer log.SetOutput(os.Stderr)
+	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Content-Type", "text/event-stream")
+		w.WriteHeader(200)
+		w.Write([]byte("data: SSE message\n\n"))
+	})
+	req := httptest.NewRequest("GET", "/events", nil)
+	recorder := httptest.NewRecorder()
+	wrappedHandler := loggingMiddleware(handler)
+	wrappedHandler.ServeHTTP(recorder, req)
+	if recorder.Code != 200 {
+		t.Errorf("Expected status code 200, got %d", recorder.Code)
+	}
+	if !strings.Contains(recorder.Body.String(), "data: SSE message") {
+		t.Errorf("Expected SSE data in response, got '%s'", recorder.Body.String())
+	}
+	logStr := logOutput.String()
+	if !strings.Contains(logStr, "WIRE_OUT SSE connection started: GET /events") {
+		t.Errorf("Expected SSE connection started log, got: %s", logStr)
+	}
+	if !strings.Contains(logStr, "WIRE_OUT SSE connection ended: GET /events") {
+		t.Errorf("Expected SSE connection ended log, got: %s", logStr)
+	}
+	if strings.Contains(logStr, "WIRE_OUT GET /events [200]:") {
+		t.Errorf("SSE endpoint should not use normal response logging, but does: %s", logStr)
+	}
+}
+
+func TestUnitChainMiddlewareOrder(t *testing.T) {
+	var order []string
+	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		order = append(order, "handler")
+	})
+	m1 := func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			order = append(order, "m1 before")
+			next.ServeHTTP(w, r)
+			order = append(order, "m1 after")
+		})
+	}
+	m2 := func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			order = append(order, "m2 before")
+			next.ServeHTTP(w, r)
+			order = append(order, "m2 after")
+		})
+	}
+	req := httptest.NewRequest("GET", "/", nil)
+	recorder := httptest.NewRecorder()
+	chainMiddleware(handler, m1, m2).ServeHTTP(recorder, req)
+	expected := []string{"m1 before", "m2 before", "handler", "m2 after", "m1 after"}
+	if len(order) != len(expected) {
+		t.Fatalf("expected %d entries, got %d", len(expected), len(order))
+	}
+	for i := range expected {
+		if order[i] != expected[i] {
+			t.Fatalf("expected order[%d]=%q, got %q (full order=%v)", i, expected[i], order[i], order)
+		}
+	}
+}
+
+func TestUnitRequireAuthRedirectsWhenUnauthenticated(t *testing.T) {
+	server := &Server{
+		authSessions: make(map[string]*AuthSession),
+	}
+	nextCalled := false
+	next := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		nextCalled = true
+	})
+	req := httptest.NewRequest("GET", "/protected", nil)
+	recorder := httptest.NewRecorder()
+	chainMiddleware(next, server.withAuth, server.requireAuth).ServeHTTP(recorder, req)
+	if nextCalled {
+		t.Fatal("expected next handler not to be called for unauthenticated request")
+	}
+	if recorder.Code != http.StatusSeeOther {
+		t.Fatalf("expected status %d, got %d", http.StatusSeeOther, recorder.Code)
+	}
+	location := recorder.Header().Get("Location")
+	if location != "/login" {
+		t.Fatalf("expected redirect to /login, got %q", location)
+	}
+}
+
+func TestUnitRequireAuthAllowsAuthenticated(t *testing.T) {
+	server := &Server{
+		authSessions: make(map[string]*AuthSession),
+	}
+	server.authSessions["token"] = &AuthSession{Email: "user@example.com"}
+	nextCalled := false
+	next := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		nextCalled = true
+	})
+	req := httptest.NewRequest("GET", "/protected", nil)
+	req.AddCookie(&http.Cookie{Name: "auth_session", Value: "token"})
+	recorder := httptest.NewRecorder()
+	chainMiddleware(next, server.withAuth, server.requireAuth).ServeHTTP(recorder, req)
+	if !nextCalled {
+		t.Fatal("expected next handler to be called for authenticated request")
+	}
+	if recorder.Code != http.StatusOK {
+		t.Fatalf("expected status %d, got %d", http.StatusOK, recorder.Code)
+	}
+}
+
+func TestUnitWithAuthSetsContextForAuthenticatedUser(t *testing.T) {
+	server := &Server{
+		authSessions: make(map[string]*AuthSession),
+	}
+	server.authSessions["token"] = &AuthSession{Email: "user@example.com"}
+
+	req := httptest.NewRequest("GET", "/", nil)
+	req.AddCookie(&http.Cookie{Name: "auth_session", Value: "token"})
+	recorder := httptest.NewRecorder()
+	handlerCalled := false
+	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		handlerCalled = true
+		ctx := authContext(r)
+		if !ctx.IsAuthenticated {
+			t.Fatal("expected request to be authenticated")
+		}
+		if ctx.Session == nil || ctx.Session.Email != "user@example.com" {
+			t.Fatalf("unexpected auth session %+v", ctx.Session)
+		}
+	})
+	chainMiddleware(handler, server.withAuth).ServeHTTP(recorder, req)
+	if !handlerCalled {
+		t.Fatal("expected handler to be called")
+	}
+}
+
+func TestUnitWithAuthSetsContextForAnonymousUser(t *testing.T) {
+	server := &Server{
+		authSessions: make(map[string]*AuthSession),
+	}
+	req := httptest.NewRequest("GET", "/", nil)
+	recorder := httptest.NewRecorder()
+	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		ctx := authContext(r)
+		if ctx.IsAuthenticated {
+			t.Fatal("expected request to be anonymous")
+		}
+		if ctx.Session != nil {
+			t.Fatalf("expected no auth session, got %+v", ctx.Session)
+		}
+	})
+	chainMiddleware(handler, server.withAuth).ServeHTTP(recorder, req)
+}
+
+func TestUnitNewLoggingResponseWriter(t *testing.T) {
+	recorder := httptest.NewRecorder()
+	lw := NewLoggingResponseWriter(recorder)
+	if lw.ResponseWriter != recorder {
+		t.Error("Expected ResponseWriter to be set to recorder")
+	}
+	if lw.statusCode != 200 {
+		t.Errorf("Expected default status code 200, got %d", lw.statusCode)
+	}
+	if lw.body == nil {
+		t.Error("Expected body buffer to be initialized")
+	}
+	if lw.body.Len() != 0 {
+		t.Errorf("Expected empty body buffer, got length %d", lw.body.Len())
+	}
+}
+
+// ---------------- Server: message parts manager and event validation ----------------
+
+func TestUnitMessagePartsOrdering(t *testing.T) {
+	manager := NewMessagePartsManager()
+	msgID := "test-message"
+	getContent := func() string {
+		parts := manager.GetParts(msgID)
+		var contents []string
+		for _, p := range parts {
+			contents = append(contents, p.Content)
+		}
+		return strings.Join(contents, " ")
+	}
+	manager.UpdatePart(msgID, "part1", MessagePartData{Type: "text", Content: "Hello"})
+	if got := getContent(); got != "Hello" {
+		t.Errorf("After first part: got %q, want %q", got, "Hello")
+	}
+	manager.UpdatePart(msgID, "part2", MessagePartData{Type: "text", Content: "world"})
+	if got := getContent(); got != "Hello world" {
+		t.Errorf("After second part: got %q, want %q", got, "Hello world")
+	}
+	manager.UpdatePart(msgID, "part1", MessagePartData{Type: "text", Content: "Hi"})
+	if got := getContent(); got != "Hi world" {
+		t.Errorf("After updating first part: got %q, want %q", got, "Hi world")
+	}
+	manager.UpdatePart(msgID, "part3", MessagePartData{Type: "text", Content: "there"})
+	if got := getContent(); got != "Hi world there" {
+		t.Errorf("After third part: got %q, want %q", got, "Hi world there")
+	}
+	manager.UpdatePart(msgID, "part2", MessagePartData{Type: "text", Content: "beautiful"})
+	if got := getContent(); got != "Hi beautiful there" {
+		t.Errorf("After updating middle part: got %q, want %q", got, "Hi beautiful there")
+	}
+}
+
+func TestUnitMessagePartsPrefixStability(t *testing.T) {
+	manager := NewMessagePartsManager()
+	msgID := "stream-message"
+	var observedPrefixes []string
+	recordContent := func() string {
+		parts := manager.GetParts(msgID)
+		var contents []string
+		for _, p := range parts {
+			contents = append(contents, p.Content)
+		}
+		content := strings.Join(contents, " ")
+		observedPrefixes = append(observedPrefixes, content)
+		return content
+	}
+	updates := []struct{ partID, content string }{{"p1", "The"}, {"p2", "quick"}, {"p1", "A"}, {"p3", "brown"}, {"p2", "very quick"}, {"p4", "fox"}, {"p5", "jumps"}}
+	expectedSequence := []string{"The", "The quick", "A quick", "A quick brown", "A very quick brown", "A very quick brown fox", "A very quick brown fox jumps"}
+	for i, update := range updates {
+		manager.UpdatePart(msgID, update.partID, MessagePartData{Type: "text", Content: update.content})
+		got := recordContent()
+		want := expectedSequence[i]
+		if got != want {
+			t.Errorf("Update %d: got %q, want %q", i+1, got, want)
+		}
+	}
+	for i := 0; i < len(observedPrefixes)-1; i++ {
+		currentPrefix := observedPrefixes[i]
+		for j := i + 1; j < len(observedPrefixes); j++ {
+			nextContent := observedPrefixes[j]
+			if !isPrefixOrValidUpdate(currentPrefix, nextContent) {
+				t.Errorf("Prefix stability violated: %q is not a valid predecessor of %q", currentPrefix, nextContent)
+			}
+		}
+	}
+}
+
+func isPrefixOrValidUpdate(old, new string) bool {
+	oldParts := strings.Split(old, " ")
+	newParts := strings.Split(new, " ")
+	if len(newParts) < len(oldParts) {
+		return false
+	}
+	for i := 0; i < len(oldParts); i++ { /* position preserved */
+	}
+	return true
+}
+
+func TestUnitMessagePartsValidation(t *testing.T) {
+	manager := NewMessagePartsManager()
+	tests := []struct {
+		name, messageID, partID string
+		wantErr                 bool
+	}{
+		{"valid inputs", "msg1", "part1", false},
+		{"empty messageID", "", "part1", true},
+		{"empty partID", "msg1", "", true},
+		{"both empty", "", "", true},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			err := manager.UpdatePart(tt.messageID, tt.partID, MessagePartData{Type: "text", Content: "test"})
+			if (err != nil) != tt.wantErr {
+				t.Errorf("UpdatePart() error = %v, wantErr %v", err, tt.wantErr)
+			}
+		})
+	}
+}
+
+func TestUnitValidateAndExtractMessagePart(t *testing.T) {
+	sessionID := "test-session"
+	tests := []struct {
+		name    string
+		event   map[string]interface{}
+		wantErr bool
+		wantMsg string
+	}{
+		{name: "valid event", event: map[string]interface{}{"type": "message.part.updated", "properties": map[string]interface{}{"part": map[string]interface{}{"sessionID": sessionID, "messageID": "msg123", "id": "part456"}}}},
+		{name: "missing messageID", event: map[string]interface{}{"type": "message.part.updated", "properties": map[string]interface{}{"part": map[string]interface{}{"sessionID": sessionID, "id": "part456"}}}, wantErr: true, wantMsg: "invalid or missing messageID"},
+		{name: "empty messageID", event: map[string]interface{}{"type": "message.part.updated", "properties": map[string]interface{}{"part": map[string]interface{}{"sessionID": sessionID, "messageID": "", "id": "part456"}}}, wantErr: true, wantMsg: "invalid or missing messageID"},
+		{name: "missing partID", event: map[string]interface{}{"type": "message.part.updated", "properties": map[string]interface{}{"part": map[string]interface{}{"sessionID": sessionID, "messageID": "msg123"}}}, wantErr: true, wantMsg: "invalid or missing partID"},
+		{name: "wrong session", event: map[string]interface{}{"type": "message.part.updated", "properties": map[string]interface{}{"part": map[string]interface{}{"sessionID": "other-session", "messageID": "msg123", "id": "part456"}}}, wantErr: true, wantMsg: "session mismatch"},
+		{name: "wrong event type", event: map[string]interface{}{"type": "other.event", "properties": map[string]interface{}{}}, wantErr: true, wantMsg: "not a message.part.updated event"},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			msgID, partID, _, err := ValidateAndExtractMessagePart(tt.event, sessionID)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("ValidateAndExtractMessagePart() error = %v, wantErr %v", err, tt.wantErr)
+			}
+			if err != nil && tt.wantMsg != "" && !strings.Contains(err.Error(), tt.wantMsg) {
+				t.Errorf("Error message %q doesn't contain %q", err.Error(), tt.wantMsg)
+			}
+			if !tt.wantErr && (msgID == "" || partID == "") {
+				t.Errorf("Valid event should return non-empty IDs, got msgID=%q, partID=%q", msgID, partID)
+			}
+		})
+	}
+}
+
+func TestUnitSequentialUpdates(t *testing.T) {
+	manager := NewMessagePartsManager()
+	msgID := "test-msg"
+	for i := 0; i < 10; i++ {
+		partID := fmt.Sprintf("part%d", i)
+		for j := 0; j < 100; j++ {
+			manager.UpdatePart(msgID, partID, MessagePartData{Type: "text", Content: fmt.Sprintf("content-%d-%d", i, j)})
+		}
+	}
+	parts := manager.GetParts(msgID)
+	if len(parts) != 10 {
+		t.Errorf("Expected 10 parts after updates, got %d", len(parts))
+	}
+	for i := 0; i < 10; i++ {
+		expectedID := fmt.Sprintf("part%d", i)
+		expectedContent := fmt.Sprintf("content-%d-99", i)
+		found := false
+		for _, p := range parts {
+			if p.PartID == expectedID {
+				found = true
+				if p.Content != expectedContent {
+					t.Errorf("Part %s has wrong content: got %q, want %q", expectedID, p.Content, expectedContent)
+				}
+				break
+			}
+		}
+		if !found {
+			t.Errorf("Missing part %s after updates", expectedID)
+		}
+	}
+}
+
+// ---------------- Server: waitForOpencodeReady helper ----------------
+
+func TestUnitWaitForOpencodeReady_DelayedSuccess(t *testing.T) {
+	var requestCount int32
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		count := atomic.AddInt32(&requestCount, 1)
+		if r.URL.Path != "/session" {
+			t.Errorf("Expected path /session, got %s", r.URL.Path)
+		}
+		if count < 3 {
+			w.WriteHeader(http.StatusInternalServerError)
+			w.Write([]byte("Server starting up..."))
+		} else {
+			w.WriteHeader(http.StatusOK)
+			w.Write([]byte(`{"sessionId": "test-session"}`))
+		}
+	}))
+	defer ts.Close()
+	var port int
+	fmt.Sscanf(ts.URL, "http://127.0.0.1:%d", &port)
+	err := waitForOpencodeReady(port, 2*time.Second)
+	if err != nil {
+		t.Errorf("Expected success after retries, got error: %v", err)
+	}
+	if finalCount := atomic.LoadInt32(&requestCount); finalCount < 3 {
+		t.Errorf("Expected at least 3 requests, got %d", finalCount)
+	}
+}
+
+func TestUnitWaitForOpencodeReady_Timeout(t *testing.T) {
+	var requestCount int32
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		atomic.AddInt32(&requestCount, 1)
+		w.WriteHeader(http.StatusInternalServerError)
+		w.Write([]byte("Permanently broken"))
+	}))
+	defer ts.Close()
+	var port int
+	fmt.Sscanf(ts.URL, "http://127.0.0.1:%d", &port)
+	start := time.Now()
+	err := waitForOpencodeReady(port, 500*time.Millisecond)
+	elapsed := time.Since(start)
+	if err == nil {
+		t.Error("Expected timeout error, got nil")
+	}
+	if err != nil {
+		expectedMsg := fmt.Sprintf("opencode server on port %d not ready after", port)
+		if len(err.Error()) < len(expectedMsg) || err.Error()[:len(expectedMsg)] != expectedMsg {
+			t.Errorf("Expected error message to start with '%s', got: %v", expectedMsg, err)
+		}
+	}
+	if finalCount := atomic.LoadInt32(&requestCount); finalCount < 3 || finalCount > 6 {
+		t.Errorf("Expected 3-6 requests in 500ms, got %d", finalCount)
+	}
+	if elapsed > 600*time.Millisecond {
+		t.Errorf("Function took too long: %v, expected ~500ms", elapsed)
+	}
+}
+
+func TestUnitWaitForOpencodeReady_ConnectionRefused(t *testing.T) {
+	unusedPort := 59999
+	start := time.Now()
+	err := waitForOpencodeReady(unusedPort, 300*time.Millisecond)
+	elapsed := time.Since(start)
+	if err == nil {
+		t.Error("Expected error for connection refused, got nil")
+	}
+	if elapsed > 400*time.Millisecond {
+		t.Errorf("Function took too long with connection refused: %v", elapsed)
+	}
+}
+
+func TestUnitWaitForOpencodeReady_JustBeforeTimeout(t *testing.T) {
+	var requestCount int32
+	successAfter := int32(4)
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		count := atomic.AddInt32(&requestCount, 1)
+		if count < successAfter {
+			w.WriteHeader(http.StatusServiceUnavailable)
+		} else {
+			w.WriteHeader(http.StatusCreated)
+		}
+	}))
+	defer ts.Close()
+	var port int
+	fmt.Sscanf(ts.URL, "http://127.0.0.1:%d", &port)
+	err := waitForOpencodeReady(port, 400*time.Millisecond)
+	if err != nil {
+		t.Errorf("Expected success just before timeout, got error: %v", err)
+	}
+	if finalCount := atomic.LoadInt32(&requestCount); finalCount != successAfter {
+		t.Errorf("Expected exactly %d requests, got %d", successAfter, finalCount)
+	}
+}
+
+func TestUnitWaitForOpencodeReady_ResponseBodyLeak(t *testing.T) {
+	bodiesClosed := int32(0)
+	requestsMade := int32(0)
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		count := atomic.AddInt32(&requestsMade, 1)
+		w.Header().Set("Connection", "close")
+		if count < 3 {
+			w.WriteHeader(http.StatusServiceUnavailable)
+			fmt.Fprintf(w, "Attempt %d failed", count)
+		} else {
+			w.WriteHeader(http.StatusOK)
+			fmt.Fprintf(w, "Success on attempt %d", count)
+		}
+		atomic.AddInt32(&bodiesClosed, 1)
+	}))
+	defer ts.Close()
+	var port int
+	fmt.Sscanf(ts.URL, "http://127.0.0.1:%d", &port)
+	if err := waitForOpencodeReady(port, 1*time.Second); err != nil {
+		t.Errorf("Unexpected error: %v", err)
+	}
+	if requests := atomic.LoadInt32(&requestsMade); requests != atomic.LoadInt32(&bodiesClosed) {
+		t.Errorf("Response body leak detected: %d requests made but only %d completed", requests, atomic.LoadInt32(&bodiesClosed))
+	}
+}
+
+func TestUnitWaitForOpencodeReady_RaceCondition(t *testing.T) {
+	var globalRequestCount int32
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		count := atomic.AddInt32(&globalRequestCount, 1)
+		if count > 5 {
+			w.WriteHeader(http.StatusOK)
+		} else {
+			w.WriteHeader(http.StatusServiceUnavailable)
+		}
+	}))
+	defer ts.Close()
+	var port int
+	fmt.Sscanf(ts.URL, "http://127.0.0.1:%d", &port)
+	done := make(chan error, 3)
+	for i := 0; i < 3; i++ {
+		go func() { done <- waitForOpencodeReady(port, 2*time.Second) }()
+	}
+	for i := 0; i < 3; i++ {
+		if err := <-done; err != nil {
+			t.Errorf("Concurrent call %d failed: %v", i, err)
+		}
+	}
+	if totalRequests := atomic.LoadInt32(&globalRequestCount); totalRequests < 6 {
+		t.Errorf("Expected at least 6 total requests from concurrent calls, got %d", totalRequests)
+	}
+}
+
+// ---------------- Race/Concurrency: UpdateRateLimiter ----------------
+
+func TestUnitUpdateRateLimiter_FirstUpdateImmediate(t *testing.T) {
+	var executed int32
+	limiter := NewUpdateRateLimiter(200 * time.Millisecond)
+	start := time.Now()
+	limiter.TryUpdate(context.Background(), func() { atomic.AddInt32(&executed, 1) })
+	time.Sleep(50 * time.Millisecond)
+	if atomic.LoadInt32(&executed) != 1 {
+		t.Error("First update should be immediate")
+	}
+	if elapsed := time.Since(start); elapsed > 100*time.Millisecond {
+		t.Errorf("First update took too long: %v", elapsed)
+	}
+}
+
+func TestUnitUpdateRateLimiter_SecondUpdateDelayed(t *testing.T) {
+	var executed int32
+	limiter := NewUpdateRateLimiter(200 * time.Millisecond)
+	limiter.TryUpdate(context.Background(), func() { atomic.AddInt32(&executed, 1) })
+	time.Sleep(50 * time.Millisecond)
+	if atomic.LoadInt32(&executed) != 1 {
+		t.Fatal("First update should have executed")
+	}
+	time.Sleep(50 * time.Millisecond)
+	limiter.TryUpdate(context.Background(), func() { atomic.AddInt32(&executed, 1) })
+	time.Sleep(50 * time.Millisecond)
+	if atomic.LoadInt32(&executed) != 1 {
+		t.Error("Second update should still be delayed at 150ms")
+	}
+	time.Sleep(80 * time.Millisecond)
+	if atomic.LoadInt32(&executed) != 2 {
+		t.Errorf("Second update should execute after interval, got %d executions", atomic.LoadInt32(&executed))
+	}
+}
+
+func TestUnitUpdateRateLimiter_UpdateAfterInterval(t *testing.T) {
+	var executed []time.Time
+	var mu sync.Mutex
+	limiter := NewUpdateRateLimiter(200 * time.Millisecond)
+	limiter.TryUpdate(context.Background(), func() { mu.Lock(); executed = append(executed, time.Now()); mu.Unlock() })
+	time.Sleep(250 * time.Millisecond)
+	limiter.TryUpdate(context.Background(), func() { mu.Lock(); executed = append(executed, time.Now()); mu.Unlock() })
+	time.Sleep(50 * time.Millisecond)
+	mu.Lock()
+	defer mu.Unlock()
+	if len(executed) != 2 {
+		t.Fatalf("Expected 2 executions, got %d", len(executed))
+	}
+	if gap := executed[1].Sub(executed[0]); gap < 240*time.Millisecond || gap > 310*time.Millisecond {
+		t.Errorf("Gap should be ~250ms, got %v", gap)
+	}
+}
+
+func TestUnitUpdateRateLimiter_RapidUpdatesCoalesce(t *testing.T) {
+	var lastValue int32
+	var executionCount int32
+	limiter := NewUpdateRateLimiter(200 * time.Millisecond)
+	limiter.TryUpdate(context.Background(), func() { atomic.StoreInt32(&lastValue, 1); atomic.AddInt32(&executionCount, 1) })
+	time.Sleep(50 * time.Millisecond)
+	if atomic.LoadInt32(&executionCount) != 1 {
+		t.Fatal("First update should have executed")
+	}
+	for i := 2; i <= 5; i++ {
+		value := int32(i)
+		limiter.TryUpdate(context.Background(), func() { atomic.StoreInt32(&lastValue, value); atomic.AddInt32(&executionCount, 1) })
+		time.Sleep(25 * time.Millisecond)
+	}
+	if atomic.LoadInt32(&executionCount) != 1 {
+		t.Error("Intermediate updates should not have executed yet")
+	}
+	time.Sleep(100 * time.Millisecond)
+	if finalValue := atomic.LoadInt32(&lastValue); finalValue != 5 {
+		t.Errorf("Expected last value 5, got %d", finalValue)
+	}
+	if finalCount := atomic.LoadInt32(&executionCount); finalCount != 2 {
+		t.Errorf("Expected 2 total executions (first + coalesced), got %d", finalCount)
+	}
+}
+
+func TestUnitUpdateRateLimiter_ConcurrentUpdates(t *testing.T) {
+	var counter int32
+	limiter := NewUpdateRateLimiter(200 * time.Millisecond)
+	var wg sync.WaitGroup
+	for i := 0; i < 100; i++ {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			limiter.TryUpdate(context.Background(), func() { atomic.AddInt32(&counter, 1) })
+		}()
+	}
+	wg.Wait()
+	time.Sleep(50 * time.Millisecond)
+	if count1 := atomic.LoadInt32(&counter); count1 != 1 {
+		t.Errorf("Expected 1 immediate execution from concurrent updates, got %d", count1)
+	}
+	time.Sleep(200 * time.Millisecond)
+	if count2 := atomic.LoadInt32(&counter); count2 != 2 {
+		t.Errorf("Expected 2 total executions after interval (immediate + coalesced), got %d", count2)
+	}
+}
+
+func TestUnitUpdateRateLimiter_TimerCancellation(t *testing.T) {
+	var executed []int
+	var mu sync.Mutex
+	limiter := NewUpdateRateLimiter(200 * time.Millisecond)
+	limiter.TryUpdate(context.Background(), func() { mu.Lock(); executed = append(executed, 1); mu.Unlock() })
+	time.Sleep(50 * time.Millisecond)
+	limiter.TryUpdate(context.Background(), func() { mu.Lock(); executed = append(executed, 2); mu.Unlock() })
+	time.Sleep(50 * time.Millisecond)
+	limiter.TryUpdate(context.Background(), func() { mu.Lock(); executed = append(executed, 3); mu.Unlock() })
+	time.Sleep(150 * time.Millisecond)
+	mu.Lock()
+	defer mu.Unlock()
+	if len(executed) != 2 {
+		t.Errorf("Expected 2 executions, got %d", len(executed))
+	}
+	if len(executed) == 2 && executed[1] != 3 {
+		t.Errorf("Expected second execution to be 3, got %d", executed[1])
+	}
+}
+
+func TestUnitUpdateRateLimiter_MultipleIntervals(t *testing.T) {
+	var executionTimes []time.Time
+	var mu sync.Mutex
+	limiter := NewUpdateRateLimiter(200 * time.Millisecond)
+	startTime := time.Now()
+	timings := []time.Duration{0, 50 * time.Millisecond, 250 * time.Millisecond, 300 * time.Millisecond, 600 * time.Millisecond}
+	for _, delay := range timings {
+		time.Sleep(delay - time.Since(startTime))
+		limiter.TryUpdate(context.Background(), func() { mu.Lock(); executionTimes = append(executionTimes, time.Now()); mu.Unlock() })
+	}
+	time.Sleep(700*time.Millisecond - time.Since(startTime))
+	mu.Lock()
+	defer mu.Unlock()
+	if len(executionTimes) != 4 {
+		t.Fatalf("Expected 4 executions, got %d", len(executionTimes))
+	}
+	tolerance := 60 * time.Millisecond
+	if executionTimes[0].Sub(startTime) > tolerance {
+		t.Errorf("First execution should be immediate, was at %v", executionTimes[0].Sub(startTime))
+	}
+	expected := 200 * time.Millisecond
+	actual := executionTimes[1].Sub(startTime)
+	if actual < expected-tolerance || actual > expected+tolerance {
+		t.Errorf("Second execution should be at ~200ms, was at %v", actual)
+	}
+	expected = 400 * time.Millisecond
+	actual = executionTimes[2].Sub(startTime)
+	if actual < expected-tolerance || actual > expected+tolerance {
+		t.Errorf("Third execution should be at ~400ms, was at %v", actual)
+	}
+	expected = 600 * time.Millisecond
+	actual = executionTimes[3].Sub(startTime)
+	if actual < expected-tolerance || actual > expected+tolerance {
+		t.Errorf("Fourth execution should be at ~600ms, was at %v", actual)
+	}
+}
+
+// ---------------- Race/Concurrency: SSE part updates and duplication ----------------
+
+func TestUnitSSEMessagePartNoDuplication(t *testing.T) {
+	manager := NewMessagePartsManager()
+	messageID := "msg_test123"
+	partID := "prt_test456"
+	if err := manager.UpdatePart(messageID, partID, MessagePartData{Type: "text", Content: "I'll analyze"}); err != nil {
+		t.Fatalf("Failed to update part: %v", err)
+	}
+	if err := manager.UpdatePart(messageID, partID, MessagePartData{Type: "text", Content: "I'll analyze OSCR's stock"}); err != nil {
+		t.Fatalf("Failed to update part: %v", err)
+	}
+	if err := manager.UpdatePart(messageID, partID, MessagePartData{Type: "text", Content: "I'll analyze OSCR's stock price over the last 6 months"}); err != nil {
+		t.Fatalf("Failed to update part: %v", err)
+	}
+	parts := manager.GetParts(messageID)
+	if len(parts) != 1 {
+		t.Errorf("Expected exactly 1 part, got %d parts", len(parts))
+	}
+	if parts[0].Content != "I'll analyze OSCR's stock price over the last 6 months" {
+		t.Errorf("Expected final content, got: %s", parts[0].Content)
+	}
+	if parts[0].PartID != partID {
+		t.Errorf("Expected partID %s, got %s", partID, parts[0].PartID)
+	}
+}
+
+func TestUnitSSEMultiplePartTypes(t *testing.T) {
+	manager := NewMessagePartsManager()
+	messageID := "msg_test789"
+	manager.UpdatePart(messageID, "prt_text1", MessagePartData{Type: "text", Content: "Analyzing data"})
+	manager.UpdatePart(messageID, "prt_tool1", MessagePartData{Type: "tool", Content: "Tool: webfetch\nStatus: running"})
+	manager.UpdatePart(messageID, "prt_tool1", MessagePartData{Type: "tool", Content: "Tool: webfetch\nStatus: completed\nOutput: ..."})
+	manager.UpdatePart(messageID, "prt_text2", MessagePartData{Type: "text", Content: "The analysis shows..."})
+	parts := manager.GetParts(messageID)
+	if len(parts) != 3 {
+		t.Errorf("Expected 3 parts, got %d", len(parts))
+	}
+	expectedOrder := []string{"prt_text1", "prt_tool1", "prt_text2"}
+	for i, part := range parts {
+		if part.PartID != expectedOrder[i] {
+			t.Errorf("Part %d: expected ID %s, got %s", i, expectedOrder[i], part.PartID)
+		}
+	}
+	if parts[1].Content != "Tool: webfetch\nStatus: completed\nOutput: ..." {
+		t.Errorf("Tool part not updated correctly: %s", parts[1].Content)
+	}
+}
+
+func TestUnitSSEHTMLGenerationNoDuplication(t *testing.T) {
+	manager := NewMessagePartsManager()
+	messageID := "msg_html_test"
+	updates := []string{"The", "The file", "The file does", "The file does not", "The file does not exist"}
+	partID := "prt_incremental"
+	for _, content := range updates {
+		manager.UpdatePart(messageID, partID, MessagePartData{Type: "text", Content: content})
+	}
+	parts := manager.GetParts(messageID)
+	if len(parts) != 1 {
+		t.Errorf("Expected 1 part after incremental updates, got %d", len(parts))
+	}
+	part := parts[0]
+	part.RenderedHTML = renderText(part.Content)
+	htmlStr := string(part.RenderedHTML)
+	if count := strings.Count(htmlStr, "The file does not exist"); count != 1 {
+		t.Errorf("Text appears %d times in rendered HTML, expected 1", count)
+		t.Logf("Rendered HTML: %s", htmlStr)
+	}
+}
+
+func TestUnitSSERapidUpdates(t *testing.T) {
+	manager := NewMessagePartsManager()
+	messageID := "msg_rapid"
+	partID := "prt_rapid"
+	done := make(chan bool)
+	go func() {
+		for i := 0; i < 100; i++ {
+			manager.UpdatePart(messageID, partID, MessagePartData{Type: "text", Content: strings.Repeat("a", i+1)})
+			time.Sleep(1 * time.Millisecond)
+		}
+		done <- true
+	}()
+	<-done
+	parts := manager.GetParts(messageID)
+	if len(parts) != 1 {
+		t.Errorf("Expected 1 part after rapid updates, got %d", len(parts))
+	}
+	if expectedLen := 100; len(parts[0].Content) != expectedLen {
+		t.Errorf("Expected content length %d, got %d", expectedLen, len(parts[0].Content))
 	}
 }
